Remove Travis debug output; work around bash bug
Apparently declaring a global array and initializing it at the same time
triggers a bug in some versions of bash 4.x prior to ~4.25 whereby it won't be
visible outside the function in which it's declared after all:

- https://lists.gnu.org/archive/html/bug-bash/2012-06/msg00068.html
- ftp://ftp.gnu.org/gnu/bash/bash-4.2-patches/bash42-025
- http://lists.gnu.org/archive/html/help-bash/2012-03/msg00078.html

Travis used bash 4.2.25(1)-release for #6 and #7, and the debug statements
for #7 clearly show that ALL_TESTS is coming up empty:

  #6:  https://travis-ci.org/mbland/go-script-bash/builds/154942364
  #7:  https://travis-ci.org/mbland/go-script-bash/builds/154943005

Locally, I'm using 4.3.46(1)-release, which presumably has the fix, which is why
I wasn't seeing it locally. Version 3.2.57(1)-release that ships with OS X
10.11.6 doesn't appear to have the problem either.

This only affects this particular test at this point, as none of the application
code uses global arrays defined in functions, and neither do any of the other
tests yet. Good to know that it apparently should be avoided to some extent,
however.