Common test environment; unset _GO_CMD
When running the tests via the shell function created by './go env -', several
tests would fail because _GO_CMD would be set to the name of the shell function,
not the name of the script. The easier thing is to unset _GO_CMD, since tests
that exercise behavior specifically dependent on it already set it directly.

Also updated one of the core-columns tests with a numeric value that serves as a
hat-tip to Jimi.