Multi-level test directories; refactor completion
This requires that for every directory containing tests, there's a 'foo.bats'
file with at least one (possibly skipped or do-nothing placeholder) test in it.
This is because of this block from bats/libexec/bats-exec-suite:

  count=0
  for filename in "$@"; do
    let count+="$(bats-exec-test -c "$filename")"
  done

Per the bash man page:

  If the last arg evaluates to 0, let returns 1; 0 is returned otherwise.

Since bats runs with the errexit option, any .bats file without any actual
test cases will cause bats to fail with no output.

The next commit will update tab completion to better support this new structure.
It still works now, but is funky because the `complete -o filename` option
causes any tab completions with a path separator in them get parsed as the last
path component, so all the nested file names are mixed in with those from higher
levels.