Revamp argument completion
The original argument completion shell function was largely inspired by rbenv's
bash completion function, which removed the word to be completed from the
argument list before passing the list through to the target script.

In reworking argument completion for the `scripts/test` script, I realized that
having access to the word being completed was critical to producing the correct
values. So this change updates the tab completion API so that the shell function
invokes `./go complete` as:

  `./go complete <word-index> [arguments]`

and `./go complete` in turn invokes the completed command as:

  `./go <command> --complete <word-index> [arguments]`

where `<word-index>` refers to one of the current command line `[arguments]`.

Note that checking $# while implementing completion may be unreliable under
certain conditions, because the shell will pass the empty string in the argument
list if the user is trying to complete an argument without typing anything
first. Therefore the argument list must/will always contain at least one
element.

This API is now documented in the command description for the `complete`
builtin.

Tests for `./go complete` are forthcoming.