path: Source in @go; remove from path, commands
Learned that adding this code to `libexec/path` won't work, because the
file-level `declare` statements are restricted to the scope of the function that
executes them. Functions, however, appear to be globally scoped regardless of
whether a function sources their implementation file.

Specifically, in the `@go` function, it sources `libexec/path` via the
`_@go.source_builtin` function. So by the time `@go` calls
`_@go.set_command_path_and_argv`, which is in the global scope, any variables
defined in `libexec/path` at file-level are already out of scope.