core: Bump minimum bash version to 3.2
Turns out 3.1 is so old it doesn't support the `extglob` and `nullglob` options.
The only place I've encountered 3.1 recently was in MinGW's MSYS distribution,
which appears really outdated and inactive. Since Bash 4.3 ships with Git for
Windows, and it's part of the Windows 10 Windows Subsystem for Linux
distribution, I'm just going to declare 3.2 the minimum rather than distort the
code to remove those options.

Note that OS X ships with 3.2 because 4.x upgraded to the GPLv3. There are some
nice-to-have features in 4.x, but none that would make a material difference to
this program.