Fix unquoted variable bugs
When running the tests in different terminals on one Windows installation,
everything passed. When running them in different terminals on another
installation, they passed in one, and failed in the others.

The reason was because on the first system, my username was one word; on the
second system, my username was two words with a space in between, and the one
terminal in which the tests passed was translating the username in my home
directory to its one-word form.