Fix unquoted variables; make tests space-sensitive
When running the existing tests in different terminals on one Windows
installation, everything passed. When running them in different terminals on
another installation, they passed in one, and failed in the others.

The reason was because on the first system, my username was one word; on the
second system, my username was two words with a space in between, and the one
terminal in which the tests passed was translating the username in my home
directory to its one-word form.

This fixes the original problem by introducing the quotes needed to get the
tests to pass in their original state. It also goes one step further by
enforcing proper quoting in more cases by having `TEST_GO_ROOTDIR` now contain
a space. Most tests broken by this only needed to have their own variables
quoted, but it also uncovered a particularly pernicious problem in
`_@go.glob_helper` from  `libexec\glob`. The problem was that the results of
the `$full_pattern` glob were getting word-split. The fix, however, was very
straightforward: just set `IFS=$'\n'`.

I also streamlined `_@go.glob_helper` slightly in the process. The insight
that `shopt -s nullglob` was unnecessary led to the removal of that option
from `lib/commands` and `libexec/complete` as well.