complete: Extract lib/complete functions
While writing tests for the original `libexec/complete`, it became
apparent that having the `_@go.complete_command_path` function in its
own module might make sense. This way it can be tested more exhaustively
in isolation, and the `help` and `path` commands can reuse it instead of
relying on the special case in `libexec/complete` for tab completion.

The next commit will replace the most of the existing
`_@go.complete_command` from `libexec/complete` with a call to
`_@go.complete_command_path`.