env: Fix test cases on MSYS2
And as fate would have it, we have to escape the closing paren as well on
MSYS2, though apparently not on the Windows Subsystem for Linux.

Also, MSYS2 didn't like the `assert_line_equals` comparison for the tab
completion function. The most straightforward thing to do seemed to be to use
`assert_output_matches` check just like with the ./go wrapper function.