assertions: Remove functions from stack trace
It's been a pain to have to mentally parse the bats stack trace each
time an assertion fails, even though it's only two lines. This finally
solves the problem by removing all functions from
`tests/assertions.bash` from the bats stack traces. The assertion tests
are slower now, since they are written to files before being executed,
but that seemed the most reliable way to validate the actual bats
output.

Turns out that `set -o functrace` is pernicious. It's not enough to set
it in a function, and then scour the stack traces. You still need to
forward to another function that does the `unset` call, otherwise
the `bats_debug_trap` gets called for `unset`, which then pushes the
function right back onto the stack again.

Also updated a couple of other ad-hoc assertion functions in other
tests in a similar fashion.