assertions: Revamp tests, fix latent bugs
The previous tests were masking the fact that some assertion failures
were getting swallowed. The problem was that I was mistakenly thinking
that `__return_from_bats_assertion` was causing its caller to return. In
some cases, it was, because it was setting `set -o errexit` originally.

I noticed the problem when writing new tests for `help` and I couldn't
get an `assert_line_equals` assertion to fail on purpose. The revamped
tests reproduced the bugs and verified their fixes.