help: Restore IFS, handle `./go commands` error
This appeared after trying to use the script with Bash 3.2.57(1)-release
(OS X 10.11.6 default) in my blog repo. The problem did not manifest
when using Bash 4.3.46(1)-release.

A seemingly related issue: apparently "temporary" assignments to `IFS`
were not meant to affect word expansions and redirections, an issue
apparently fixed in Bash 4.3:

https://lists.gnu.org/archive/html/bug-bash/2014-03/msg00065.html

Perhaps the "fix" prevents called functions from inheriting the value of
`IFS`. Certainly adheres to the principle of least surprise considering
the ubiquity and importance of `IFS`, but violates the principle of
expecting called functions to inherit variable values from their
callers. In a way, both versions were doing the "right" thing, but in
the case of 3.2, the "right" thing exposed a latent bug.