glob: Fix GLOBIGNORE test for bash 4.4, add note
Bash 4.4 was released last night, and apparently the `GLOBIGNORE`
behavior changed in a fashion that broke the 'glob/main: recursively
discover files' case. Specifically, a pattern with a leading `*` doesn't
seem to match arbitrarily deep subdirectory prefixes anymore. Expanding
the `--ignore` argument to be more specific fixes the test for bash 4.4
and still works for previous versions.