command_descriptions: Trim trailing whitespace
This only really is an issue when the file only contains one line
without a newline--such as in dummy test files--because the loop never
executes to remove the whitespace appended to the end. But it seems
worth fixing at this level rather than leaving myself and others
scratching our heads when test output spits out an otherwise mysterious
whitespace.

Also removes some superfluous `assert_success` invocations.