core: Export _GO_CMD_{NAME,ARGV}, add Bash test
Theoretically, any command can now re-invoke the original `./go` script
command via:

  "$_GO_SCRIPT" "${_GO_CMD_NAME[@]}" "${_GO_CMD_ARGV[@]}"

In scripts written in other languages, the `_GO_CMD_NAME` and
`_GO_CMD_ARGV` environment variables will have to be split using ASCII
NUL as the delimiter to get the arguments, but the idea is the same.

`_GO_CMD_NAME` will be used to implement the generic
`lib/test` module in a future commit. Specifically, it will enable
`@go.test_coverage` to invoke `kcov` using the original command
arguments without requiring the caller to pass the name of the test
command script as an argument, or requiring that it always be named
`test`.