complete: Replace most compgen calls with echo
Per the previous commit, since the results from running a script with
`--complete` filter through `compgen -W` in `libexec/complete` anyway,
in most cases it seems better to echo potential matches and let
`libexec/complete` do the work.

In a couple places, we retain `compgen -f` and `compgen -d`, as these
are trickier to replace with feeding `echo` the appropriate globs.