complete: Eliminate compgen from internal library
Since the output was always fed to `compgen -W` in _@go.complete_args to
begin with, this was logically a no-op for most purposes. However, the
function now only returns success if the caller should execute the
command script specified by `__go_cmd_path`, and the
complete/command-path test required a bit of refactoring.