log: Add @go.log_command, critical section flag
This enables scripts to both execute and log commands in a single line,
recording the command arguments and the exit status if unsuccessful.

The critical section flag allows scripts to abort with a `@go.log FATAL`
if any one of a series of commands fails. This facility is reentrant.

There's also a `_GO_DRY_RUN` flag that, when set, will cause
`@go.log_command` to only log the command arguments without executing
the command.