core: Update how COLUMNS is set
Using `C:\Program Files\Git\bin\bash.exe` from a vanilla Command Prompt window
set to 161 columns, the `default to 80 columns if tput fails` test case kept
failing because columns was getting computed as 161 instead of 80. Eventually
I realized this wasn't because the test wasn't getting `tput` to fail--it
was--but because `mode.com` was picking up the slack, as designed.

So the test has been updated to `default to 80 columns if tput fails or use
mode.com on Windows`. Also streamlined the existing `go-core.bash` logic that
handles `tput` and `mode.com` to use regular expressions and `BASH_REMATCH`,
dispensing with `shopt` and multiple steps in the `mode.com` case.