core: Reproduce and fix `tput cols` error from #17
Turns out that redirecting standard error to `/dev/null` has the effect
of causing `tput cols` to output '80', which defeats the entire purpose
of trying to set `COLUMNS` dynamically with `tput` in the first place.

To reproduce the problem in the test, I had to realize that Bats will
redirect standard output and standard error to a file, and override that
behavior by redirecting both back to `/dev/tty`--which Bash conveniently
emulates on Windows as well. Then I had to go back to the `command -v`
checks for `tput` and `mode.com`, though I kept the `BASH_REMATCH`
assignment in the `mode.com` case.