Revamp `_GO_*` var exports, add `vars` builtin
Now the `_GO_*` variables are easy to discover and examine, and their
contract is well-documented, thanks to the new `vars` builtin command.

Of course the irony is that the implementation is so trivial, yet the
documentation and testing is so extensive, because it has to be.

Also, the `_GO_*` variable tests from
`tests/core/run-command-scripts.bats` have been moved to
`tests/vars.bats`.