use: Improve module import error message
It could be that the module has syntax errors, or that it failed to
define or import all its necessary functions and variables, or that it
deliberately raised an error based on an error condition. It's better to
state plainly that the import failed, and permit the user to ascertain
the cause from the module's output, rather than making it sound like the
module script is broken.