log: Add optional timestamp prefix to messages
Closes #34. Tested on macOS 10.12.2 under bash 4.4.5(1) from Homebrew
and the bash 3.2.57(1) that ships with the system, and confirmed that
the right tests execute and pass, and that the right tests are skipped
for each version.

Required the envisioned assignment of `__GO_LOG_TIMESTAMP_IMPL` to take
place in `_@go.log_init` instead, as the `WARN` message in the case of
unavailable timestamps was emitted into the first `@go.log` call
otherwise (caught by the "log/timestamp: log messages prefixed with
timestamp if supported" test case when testing with bash 3.2.57(1)).