modules: Switch from `declare` to readonly, export
While writing the upcoming `@go.log_add_output_file` function in the
`log` module, which includes a call to `. "$_GO_USE_MODULES" 'file'`, a
test that made multiple calls to this new function would fail with the
message:

 ✗ log/add-output-file: add output files for a mix of levels
   (in test file tests/log/add-output-file.bats, line 83)
     `"@go.log_add_output_file '$TEST_GO_ROOTDIR/info.log' 'INFO'" \' failed
   line 0 not equal to expected value:
     expected: 'INFO   FYI'
     actual:   'No file descriptors <  available; failed at:'

This error message is part of `@go.open_file_or_duplicate_fd`, and
includes a reference to `$_GO_MAX_FILE_DESCRIPTORS`, which was
previously declared as:

  declare -r _GO_MAX_FILE_DESCRIPTORS="${_GO_MAX_FILE_DESCRIPTORS:-256}"

The fact that no value appeared for "$_GO_MAX_FILE_DESCRIPTORS" was
suspicous. Moving the `. "$_GO_USE_MODULES" 'file'` call outside of the
function to the top level of the module got the test to pass; I
understood this to be because:

- Top-level `declare` statements cause variables to be scoped to the
  function sourcing the file.
- When the function call ends, the declared variables go out of scope.
- Since "$_GO_USE_MODULES" ensures module files are imported only once,
  these variables aren't redeclared for subsequent calls.

After deciding that it should be permissible for a function to call
`.  "$_GO_USE_MODULES"` and have it work across multiple calls, I
eventually learned that substituing `readonly` for `declare -r` did what
I originally intended. The same also holds for `export` vs `declare` or
`declare -x`. Making these updates to the `file` module allowed the test
to pass; only minor changes were required to other tests when applying
the updates to every module.

For now, this update applies only to module code. I may look into how to
apply `readonly` and `export` to other code as well, but an initial
experiment with replacing values in `go-core.bash` caused `./go test
vars` to fail.