stack-trace-item: Allow function opening as needle
This is in anticipation of the following commit to update
`@go.log_command`, which contains a process redirection.

For some reason, stack traces from commands executed within a process
redirection will show the caller's line as the opening of the function
definition, rather than the line of the process redirection.

If the invocation is on a line of the process redirection other than the
first, then the reported line will be offset from the opening of the
function definition. If it becomes necessary, we can add an offset
parameter to account for this later.