validation: Add identifier validation, fatal funcs
It seems the more common validation case I need so far is for validating
that callers are passing in the correct variable name for an `eval
"$var=\$value"` assignment. It also seemed like a good idea to include
fatal versions of these functions, since in many cases it is a
programmer error when validation fails and the program should crash
loudly.