log: Add `log/log-command` test for whitespace
The intention was to refactor `@go.log_command` to hoist the temporary
`IFS=` assignment out of the `while IFS= read` loop by demonstrating how
this deeply nested series of `@go.log_command` invocations would fail
under Bash 4.2.25. However, the test continued to pass.

During the course of experimentation, I tried reverting
`_@go.log_level_file_descriptors` back to setting `IFS=, read` inline as
before #81, and the new test failed as expected, only producing the
output:

```
   RUN    outer_function
      Hello, World!
```

Then I replaced the implementation with this, and the test passed:

```
  _@go.log_level_file_descriptors() {
    while IFS=',' read -ra __go_log_level_file_descriptors; do
      break
    done <<< "${__GO_LOG_LEVELS_FILE_DESCRIPTORS[$1]}"
  }
```

So it appears that it's only single `IFS= read` statements that are
vulnerable to the bug, but `while IFS= read` statements are fine. Good
to know.