tests: Replace `[[ ]]` with `[ ]` because Bash 3.2
While writing test cases for `lib/bats/assertion-test-helpers` per #49,
I noticed that some failing `[[ ]]` conditions were causing the stack
trace to point to the previous line—if it was the last line of the test
case. I eventually noticed that `[[ ]]` conditions earlier in a test
case would always pass regardless of what was getting checked, even when
`echo "$?" >&2` on the following line would report the expected status.

That was when running the test with Bash 3.2.57(1)-release on macOS.
When I switched to Bash 4.4.5, the `[[ ]]` conditions behaved as
expected. I reasoned that somehow `[[ ]]` wasn't triggering the ERR trap
like `[` was, perhaps because `/bin/[` is an external command. Then I
found this in https://tiswww.case.edu/php/chet/bash/CHANGES:

  This document details the changes between this version,
  bash-4.1-alpha, and the previous version, bash-4.0-release.

  j.  The [[ and (( commands are now subject to the setting of `set -e'
      and the ERR trap.

Yeah. So I was right. On the bright side, under Bash 3.2.57(1)-release,
`command -V '['` reports "[ is a shell builtin", so there shouln't be a
performance penalty. But it's still frustrating that we can't use the
safer `[[ ]]` construct in portable Bats test cases.