ass-test: Add test cases for missing conditions
Noticed the missing `__check_expected_output` test paths thanks to:
https://coveralls.io/builds/9590074/source?filename=lib%2Fbats%2Fassertion-test-helpers

Also realized that I needed to add tests to ensure the
`expect_assertion_*` functions catch the error when
`return_from_bats_assertion` is delegated. It's essentially the same
condition as when the call is skipped altogether, but it's good to check
for both conditions explicitly.