lib/testing: Fix stack-trace-item regex on Linux
The following Travis builds for #105 failed, when everything worked
locally on macOS 10.12.2 fine on my machine across Bash 3.2.57, 4.2.25,
and 4.4.5:

https://travis-ci.org/mbland/go-script-bash/jobs/191473802
https://travis-ci.org/mbland/go-script-bash/jobs/191473815

It worked on Windows 10, too. But on both Ubuntu and Arch Linux, the
same tests failed in the same way, due to an incompatibility in the
`function_opening` regex. Turns out some systems are more forgiving of
an unescaped `{` than others, as it typically is the beginning of a
bound (one or more integers specifying the number of times the
preceding pattern is repeated).

The fix was just to escape the `{`. Escaped the `()` in the same way.