lib/log: Refactor _@go.log_format_level_labels
The updates to `lib/format` from #107 make this function much shorter.
Also, the logic to check the file descriptors was sort of confused and
not really doing anything, at least not anything helpful, since the
decision to print or strip formatting codes is now made per log level
file descriptor in `@go.log`.