tests: Replace Bats submodule with a shallow clone
The theory is that this will work better when trying to compose projects
using the framework and its plugins, since it won't be a web of
submodules all the way down. By creating shallow clones on demand when
the `./go` script for the framework, a plugin, or an application is
executed, we may avoid creating circular dependencies and generally
reduce the amount of bandwidth and disk space consumed by each build.