path: Remove _GO_PLUGINS_DIR from _GO_SEARCH_PATHS
Part of #120. Now scripts that may be exported as plugins must reside in
a project's `/bin` directory, not the root directory. Some implications
of this:

The top-level `./go` scripts in plugins will not appear in the parent's
available commands. This way a plugin can have its own `./go` script for
development.

Applications intended to function both standalone and as a plugin are
possible. There may be separate `./go` scripts and script directories
for the standalone application interface and for the development
interface (e.g. `./dev` with its own scripts dir, which could possibly
even be `bin/dev`).

Applications that want to export plugins, but not the entire
application, must extract a new project repository that is imported into
the original application. In addition to freeing potential clients from
the original application's dependencies, it's a forcing function to
ensure the original application's team keeps the plugin interface
up-to-date.

This change also ensures `_GO_PLUGINS_DIR` is always set. This will have
implications in a future commit such that for scripts running as a
plugin, this will be the plugin dir for the top-most script.

A bunch of the changes in this commit are to the
`tests/commands/helpers.bats` functions and how they're called from the
`commands/` tests.