path: Search `_GO_SCRIPTS_DIR` before plugin dirs
Part of #120, and similar to #122, searching the local `_GO_SCRIPT_DIR`
before plugin dirs seems like a more natural and scalable model.
`_GO_CORE_DIR/libexec` for builtins still comes first.

The idea behind this model is that when reading a script within the
`./go` script framework, it should be clear that:

- core framework scripts are the same everywhere; they can't be shadowed
  by scripts with the same name
- project-local scripts take precedence before any plugin scripts

The implications of this last point are that:

- a script that is installed as a plugin will find its own scripts
  before those that share the same name with other installed plugins,
  avoiding collisions and possibly obscure bugs
- the parent project can determine the precedence of plugin scripts
  sharing the same name, as the plugins can be installed using directory
  names that force a particular order (e.g. '00-', '01-', '02-'
  prefixes).
- as an alternative to the approach in the preceding point, the parent
  project can potentially resolve conflicts by installing its own
  scripts that then delegate to a specific plugin script

And to this last point, I've filed #127 to facilitate help text
discovery.