core, path: Refactor plugin path handling
I realized the `while true` loop from `_@go.run_plugin_command_script`
could live comfortably in `_@go.set_search_paths` and handle the
top-level `_GO_PLUGINS_PATHS` case as well.

Also, I realized that there was no need for `@go` to invoke
`_@go.run_plugin_command_script` if the command script in question was
from the `_GO_SCRIPTS_DIR` of the current plugin and not in
`_GO_SCRIPTS_DIR/plugins`.

Plus, there was a bug whereby it'd been presumed that the plugin command
path would always reside in the top-level `_GO_SCRIPTS_DIR` of the
plugin. Now the plugin's `_GO_SCRIPTS_DIR` is correctly set to the
`/bin` parent directory of the command script. I'll add thorough tests
for this in a future commit/PR.