modules: Ensure plugin's modules precede parent's
Realized that if a plugin contains an internal or exported module that
matches the same filename as an internal or exported module in the
parent, that the code as it was would find the parent module first. The
fix was straightforward: extract most operations into
`_@go.find_module`.

Touching up the "modules/use: prevent self, circular, and multiple
importing" test case required a bit more care to touch-up, since the
plugin module in the test could no longer access its parent's modules
any longer.