testing/env: Make `@go.test_compgen` fail loudly
The previous commit rectified errors in assertions from `tests/new.bats`
whereby assertions that should have failed as written did not. Though
the code under test was sound, these latent test errors were due to the
fact that `@go.test_compgen` would produce an empty result array and
return success if the underlying `@go.compgen` returned empty results.

When this empty array was passed to `assert_success` or `assert_failure`
using "${array[@]}" notation, it expanded to an empty argument list, in
which case these assertions did not check the value of `$output`. This
left the impression that the output matched a nonempty list of values
when the assertions were actually not comparing the output at all.

Now `@go.test_compgen` will fail loudly with `@go.compgen` results are
empty, and any case in which the underlying `@go.compgen` returns an
error.