use/plugins: Refactor test to use _G_U_M, not @go
This is part of the refactoring of `_GO_USE_MODULES` to provide correct
`@go` command script scope for plugin modules.

Specifically, this makes sure that the module search paths are set
correctly independent of the `@go` function being called. This is to
ensure that modules behave as expected when imported before the
top-level `./go` script calls `@go`.