core: Preserve true line endings in @go.printf
Closes #146. Works by adding the ASCII Unit Separator character after
every newline character and using it as the delimiter for `read`. This
way output that shouldn't end with a newline won't.

Seems to add about 0.003-0.005s to the time on average as measured by:

  time COLUMNS=<N> ./go help builtins

regardless of the value of `<N>` used to set `COLUMNS`. Also measurably
faster than the alternate methods of:

- counting the total number of characters printed and only emitting a
  newline if the number of `result` characters hasn't been reached
- counting the number of lines and only emitting a newline if the
  current line number is less than or equal to the length of
  `${result//[^$'\n']}`

Also includes an update to `libexec/get.d/file` to fix a test broken by
this new behavior, as it previously depended upon the old behavior to
add a newline by default.