core: Fix latent @go.printf infinite loop bug
Noticed the bug when trying to run:

  time COLUMNS=27 ./go help builtins

The script would loop infinitely when trying to print the line for the
`path` builtin. When the `@go.printf` loop would reach the point where
the `prefix` variable contained:

  'script, [alias] or'

the expression:

  line="${line#$prefix}"

would cause `[alias]` to be interpreted as a set of characters to match
per standard Bash pattern matching, rather than as a literal string in
its own right. Consequently, the prefix would not get removed from
`line`, so `line` would remain constant and the loop condition
`"${#line}" -gt "$COLUMNS"` would never fail.

The fix was to use index notation and the length of `prefix` to trim
`line` instead:

  line="${line:${#prefix}}"

The included test case reproduces the bug and verifies its fix.