template: Fix Windows/MSYS2 test failure
On MSYS2, `/tmp` (used in `TEST_GO_ROOTDIR` via `BATS_TMPDIR`) is mapped
to `C:/Users/<user>/AppData/Local/Temp/`. When `git clone ... ` is called
with a path starting with `/tmp`, its output show the `C:/Users/...`
path instead.

The fix is to update the `assert_output_matches` expression to match
only the `scripts/go-script-bash` suffix of the clone path instead of
the entire path.