testing/env: Add `@go.test_compgen`
Regarding the claim that this new function is faster than the old
hand-written loops, here are representative timings for the command:

  TEST_FILTER='(directories|: tab completion$)' gos test complete get/

For hand-written loops:

  43 tests, 0 failures, 39 skipped

  real    0m7.020s
  user    0m3.352s
  sys     0m3.473s

For `@go.test_compgen`:

  43 tests, 0 failures, 39 skipped

  real    0m6.367s
  user    0m3.071s
  sys     0m3.104s

Considering the fact that most of that time was spent skipping 39 tests,
the result is pretty substantial. Even beyond that, it provided the
insight that led to filing issue #156.