cmd-desc: Make `from-file` test faster
The reason this test suite was so slow before was because the test cases
made direct calls to the functions under test instead of wrapping them
inside `TEST_GO_SCRIPT`. The functions would then call other functions,
then call other functions in loops, and so on, causing Bats to do a lot
of work building stack traces with every call.

A sample time before this change on my macOS MacBoook Pro:

  6 tests, 0 failures

  real    0m6.554s
  user    0m2.820s
  sys     0m3.487s

A sample time after:

  6 tests, 0 failures

  real    0m2.316s
  user    0m1.010s
  sys     0m1.204s

I expect the difference to be even more dramatic on Windows.