lib/bats: Extract `lib/bats/helper-function`
First part of #156. Though it's still a work in progress, the speedup in
the tests with this new behavior added to the functions from
`lib/testing/environment` is substantial enough that I'm eager to get it
merged sooner than later. The other changes will be more extensive, so
this will reduce the amount of noise in the pull requests as well.

Also, I finally grokked Bats enough to understand why processing
`BATS_PREVIOUS_STACK_TRACE` is unnecessary. I've removed that
processing everywhere, and added a thorough comment to
`restore_bats_shell_options`.