lib/test/stack: Use `lib/bats/helper-function`
Per the deleted comment from `@go.stack_trace_item`, I've figured out
why it was faster to run `for` and `while` loops as a new process. Since
Bats registers the `bats_debug_trap` function to collect stack trace
information, and `DEBUG` traps fire before every iteration of a loop,
that cost slowed down the implementation when it was originally
in-process.

Now that the `set "$DISABLE_BATS_SHELL_OPTIONS"` and
`restore_bats_shell_options` mechanism disables `DEBUG` traps from being
inherited by invoked functions, we can add `__go.stack_trace_item` back
to the `lib/testing/stack-trace` file and remove the external
`lib/testing/stack-trace-item` script. This commit applies the mechanism
to other public API functions as well.

This also results in a speed increase on my MacBook Pro for the
following command:

  ./go test file/open-or-dup log/add-output-file log/add-update \
    log/filters log/log-command log/main log/setup-project modules/use \
    testing/stack-trace

from this:

  119 tests, 0 failures

  real    0m37.419s
  user    0m18.710s
  sys     0m16.910s

to this:

  119 tests, 0 failures

  real    0m31.346s
  user    0m16.186s
  sys     0m13.627s

The speedups from #159 and #160 already pulled test times on Windows
from the high 50-minute range down to the mid 30-minute range, so I look
forward to the impact this change will have on that platform.